move BillingEntry to billing
