+++ fixbackend/billing/models.py
@@ -1,4 +1,4 @@
-#  Copyright (c) 2023. Some Engineering
+#  Copyright (c) 2023-2024. Some Engineering
 #  This program is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU Affero General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -11,11 +11,22 @@
 #
 #  You should have received a copy of the GNU Affero General Public License
 #  along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-
+from datetime import datetime
 from typing import Union, List
-from fixbackend.ids import SubscriptionId
+from fixbackend.ids import SubscriptionId, BillingId, WorkspaceId, ProductTier
 from attrs import frozen
+
+
+@frozen
+class BillingEntry:
+    id: BillingId
+    workspace_id: WorkspaceId
+    subscription_id: SubscriptionId
+    tier: ProductTier
+    nr_of_accounts_charged: int
+    period_start: datetime
+    period_end: datetime
+    reported: bool
 
 
 class PaymentMethods:
