+++ fixbackend/billing/schemas.py
@@ -1,4 +1,4 @@
-#  Copyright (c) 2023. Some Engineering
+#  Copyright (c) 2023-2024. Some Engineering
 #  This program is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU Affero General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -17,11 +17,10 @@ from datetime import datetime
 from pydantic import BaseModel, Field
 
 from typing import List, Literal, Optional, Union
-from fixbackend.billing.models import PaymentMethods, WorkspacePaymentMethods
+from fixbackend.billing.models import PaymentMethods, WorkspacePaymentMethods, BillingEntry
 from fixbackend.billing import models
 
 from fixbackend.ids import BillingId, ProductTier, SubscriptionId, WorkspaceId
-from fixbackend.subscription.models import BillingEntry
 from enum import Enum
 
 from fixbackend.workspaces.models import Workspace
